+++ changelog.md
@@ -7,7 +7,7 @@
 - Avoid where possibile to recreate already existent object. Now i try to recover from the canvas/scene/game every object i needed for the trigger instead to create runtime from a wrong id or name, i lose something in performance but i think is the right road to follow
 - Add gulp build file for more relaxing build of the source and developing
 - Add github action for generate release with 'module.zip' and 'module.json'
-- Add integration with Stairways mdule
+- Add integration with Stairways module, limited only to click :(
 - Add integration with Tagger mdule (details on readme)
 - Made check for entity case insensitive avoid many minor issue form miss typinf the name ecc.
 - Add the setting 'Enable tagger integration feature' (details on readme)